POCOR-4617
Academic Period and Committee Type filters are added into InstitutionCommittee index.
InstitutionCommitteeTabsComponent removed as the default page behaviour, overrides the querystring when the academic period and committee type filters are selected,
hence tab elements are set in each individual controllers instead.
This is to prevent the error of setting invalid links to the tabs.