POCOR-4679
Disable numeric validation if contact types validation pattern is not null.
This because the validation pattern(regex) will check the given user input and if the validation pattern allows user to enter special characters, user will not be able to save because of the numeric validation.
User have to specify a regex pattern with the special characters in Field Options > Contact Types to allow user to enter special characters in the contact input.