POCOR-4685
Migration added to insert a new column next_institution_class_id to institution_class_students
Added a new column Next Class and a dropdown.
The values in the Next Class dropdown are populated based on the selected academic period and education grade.
If the Student Status selected is GRADUATED, there will be no options for user to select.
New button "Save As Draft" added to the reconfirm page and only the next_institution_class_id is saved when clicked
Student status id and student will also be added to the new class
New message added for successful saving of draft.