POCOR-4919 Use another variable for the appended excel url with academic period id so that the academic period id will not be appended each time user changes the academic period. Added StaffAttendancesTable for generating the attendance excel. Logic for attendance type
PRESENT: As long as user has time in without any leave
HALF DAY ABSENT: If leave record is found and the full_day is 0. If attendance record found for the same day, the time in and time out record will also be printer
FULL DAY ABSENT: If leave record is found and the full_day is 1.
CLASS ATTENDANCE NOT MARKED: If both leave and attendance records is not found for a day.